'arc' and 'semicircle' commands implemented
 - Arc class fleshed out
 - Point#angle(a, b) implemented
 - Register#remove(category, shape) implemented
 - Shape#deregister() implemented

Removing shapes from the register is necessary because creating an arc
involves creating a circle (for construction purposes only) which must
not go in the register.  However, this is a hack.  Future work will
prevent second-class shapes from getting into the register in the first
place.

Arcs based on sloping diameters or radii do not currently render correctly.

Finally, added NEWS file to track development.