Progress in implementing polygons
Enough fixed to the code have been made that two of the tests FAIL
instead of ERROR.  This is significant, as it means a lot of missing
methods were implemented and other stuff fixed up.  Among the changes:

 * In a shape declaration, the 'label' property can now be a string
   like "3+", meaning the label must have three or more characters.

 * Label#points and Label#mask -- labels are smarter than before.

 * Bug in Point.polar(r,t) fixed.

 * Polygon-awareness added to the requires in rgeom.rb and the register.
   This manual maintenance has its benefits but it's a bit annoying.

 * Circle#angle(theta) -> point   and   Circle#angle_at(point) -> theta

 * Bug in isosceles triangles found and fixed on master.
   - Tests added in this commit

Obviously a lot more to go: none of the tests pass yet.